+++ Clicked/Core/BindingProcessor.lua
@@ -1026,23 +1026,13 @@ function Addon:GetMacroForBindings(bindings, interactionType)
 
 				if interactionType == Addon.InteractionType.REGULAR then
 					if not startAutoAttack and binding.action.startAutoAttack then
-						for _, target in ipairs(binding.targets.regular) do
-							if target.unit == Addon.TargetUnits.TARGET and target.hostility ~= Addon.TargetHostility.HELP then
-								startAutoAttack = true
-								table.insert(lines, "/startattack [@target,harm]")
-								break
-							end
-						end
+						startAutoAttack = true
+						table.insert(lines, "/startattack")
 					end
 
 					if not startPetAttack and binding.action.startPetAttack then
-						for _, target in ipairs(binding.targets.regular) do
-							if target.unit == Addon.TargetUnits.TARGET and target.hostility ~= Addon.TargetHostility.HELP then
-								startPetAttack = true
-								table.insert(lines, "/petattack [@target,harm]")
-								break
-							end
-						end
+						startPetAttack = true
+						table.insert(lines, "/petattack [@target,harm]")
 					end
 				end